requirements.txt
streamlit
openai
opencv-python
python-dotenv
pillow
requests